Capable of picking a dog to filter on
Restructured how the info gets pulled out

Need to decide how to save down the filters
Reconstructured Database

Left off at dogPreferences for the filter, the ID gets pulled out as do the preferences (which I now realise needs to be changed to personality,a dn prefences too), anyway. The dog name needs to be added, create overview item
Stage?


Capable of picking a dog to filter on


Changes on filter